add 3190 header file : snake.h
